Starting JSON

+++ SteamCMDHelper.py
@@ -4,28 +4,40 @@
 import sys
 import os
 import argparse
-# import json
+import json
 # import subprocess
 from zipfile import ZipFile
 
 import urllib.request
-# import urllib.parse as url_parse
 
+class Settings:
+    install_dir = ""
+    installed_games =	{
+        "example_game": 123 #Once im done working on this i will command this line out
+    }
+
+
+
+#Would put this in an enum but I cant think of a good reason to, its only 3 levels
 level_all = 0
 level_verbose = 1
 level_debug = 2
-# log_level = 0
 
 parser = argparse.ArgumentParser(description="Python based Tool for using the SteamCMD")
-# parser.add_argument('-d', '--dry-run', action='store_true', dest='dry_run', help="Dont execute downloads or modify files, just prints what would happen") #todo, all the things
+parser.add_argument('-dr', '--dry-run', action='store_true', dest='dry_run', help="Dont execute downloads or modify files, just prints what would happen") #todo, all the things
 parser.add_argument('-v', '--verbose', action='store_true', dest='verbose', help="Enable verbose logging")
 parser.add_argument('-D', '--debug', action='store_true', dest='debug', help="Enable debug logging. This should only be used if there are problems.")
 
 args = parser.parse_args()
 def main():
-    #GLOBAL STUFF
+    #GLOBAL VARIABLES
+    
+    #global logging level
     global log_level
     log_level = 0
+    #global dry run flag
+    global dryrun
+    dryrun = False
 
     #END GLOBAL
     if sys.version_info[0] < 3:
@@ -40,10 +52,24 @@ def main():
         log_level = 2
         d_print(level_debug, "Debug Logging Enabled")
 
+    if args.dryrun:
+        d_print(level_all, "Dry Run Enabled - No files will be downloaded and no commands will be passed to the SteamCommandLine")
+        dryrun = True
+
+
+    #USER SETTINGS JSON - not sure how large this will end up being
+    #But for now itll at least be steamcmd install location and a list of the games installed so far
+
 
     d_print(level_all, "\n---------------------------\n")
     d_print(level_all, "Welcome to the SeamCMDHelper")
     d_print(level_all, "\n---------------------------\n")
+
+
+    #MAIN USER INPUT LOOP
+
+
+
 
     
     #download
@@ -51,6 +77,17 @@ def main():
 
     #print(sys.platform)
     return 0
+
+def LoadJSON():
+    import json #why do i have to do this here?
+    d_print(level_debug, "Loading UserSettings JSON")
+    json = json.loads(open("UserSettings.json"))
+    userSettings = Settings()
+    return 
+
+def SaveJSON():
+    d_print(level_debug, "Saving UserSettings JSON")
+
 
 def downloadCMD(install_location):
     #always make the folder, if it already exists I doubt itll fail
@@ -63,12 +100,14 @@ def downloadCMD(install_location):
     if sys.platform == 'win32' or sys.platform == 'win64':
         d_print(level_debug, "I THINK I AM WINDOWS, SO I WILL DOWNLOAD A ZIP")
         url = "https://steamcdn-a.akamaihd.net/client/installer/steamcmd.zip"
-        urllib.request.urlretrieve(url, install_location+"steamcmd.zip")
+        if(not dryrun):
+            urllib.request.urlretrieve(url, install_location+"steamcmd.zip")
 
     if sys.platform == 'linux':
         d_print(level_debug, "I THINK I AM LINUX, SO IM GOING TO TRY TO DOWNLOAD A TAR")
         url = "https://steamcdn-a.akamaihd.net/client/installer/steamcmd_linux.tar.gz"
-        urllib.request.urlretrieve(url, install_location+"steamcmd_linux.tar.gz")
+        if(not dryrun):
+            urllib.request.urlretrieve(url, install_location+"steamcmd_linux.tar.gz")
 
     
 def d_print(level, text):
